Use em instead of px in special link
Why? This allows us to keep the exact same proportional distances even
if we change the font size in other components.

This is super powerful: my rythm is always the same, whatever the font
size. With pixels, it would be completely destroyed outside of the exact
situation where I designed the layout first.

I finally get it. Fiou. 15 years after starting webdev. Well, at least,
I was using em even if the purpose was a bit fuzzy.